refactor components -> directories
